StatusCodes & MediaTypes in objects
